Fix a bug in WorkflowTest.java
The WorkflowTest rule does not use the DataConverter that is specified in the Workflow annotation. Instead, it always uses the JsonDataConverter. This means that if you try to use your own DataConverter with the WorkflowTest rule, your tests will fail if the JsonDataConverter can't deserialize something that your custom DataConverter serialized.

This change creates an overloaded constructor that will allow developers to construct the WorkflowTest rule with their custom DataConverter implementation.